eidting product and product catagory test files

+++ src/features/Products/productService.js
@@ -12,6 +12,10 @@ exports.addCatagory = async (req)=>{
     if(catagory){
         throw new Error('The product catagory already exists')
     }
+    const user = await userModel.User.findOne({email:req.user.email})
+    if(user.role != 'admin'){
+        throw new Error("permission denied: only admin have access to catagory")
+    }
     const newcatagory = new productModel.productCatagory({
         name:req.body.name,
         description:req.body.description
@@ -21,6 +25,14 @@ exports.addCatagory = async (req)=>{
 }
 
 exports.getCatagory = async (req)=>{
+    const user = await userModel.User.findOne({email:req.user.email})
+    
+    const role = ['Seller','admin']
+    if(!role.includes(user.role)){
+        console.log(!role.includes(user.role))
+        throw new Error("permission denied: only admin and seller to get catagory")
+    }
+
     const catagory = await productModel.productCatagory.findOne({name:req.params.name})
     if(!catagory){
         throw new Error(`${req.params.name} is not found`)
@@ -36,6 +48,12 @@ exports.updateCatagory = async(req)=>{
         const errorMessages = err.array().map(error=>error.msg).join(' ')
         throw new Error(errorMessages)
     }
+
+    const user = await userModel.User.findOne({email:req.user.email})
+    if(user.role != 'admin'){
+        throw new Error("permission denied: only admin have access to catagory")
+    }
+
     const catagory = await productModel.productCatagory.findOneAndUpdate(
         {name:req.params.name},
         {name:req.body.name,description:req.body.description},
@@ -44,10 +62,15 @@ exports.updateCatagory = async(req)=>{
     if(!catagory){
         throw new Error(`${req.params.name} is not found`)
     }
+    
     return catagory
 }
 
 exports.deleteCatagory = async (req)=>{
+    const user = await userModel.User.findOne({email:req.user.email})
+    if(user.role != 'admin'){
+        throw new Error("permission denied: only admin have access to catagory")
+    }
     const catagory = await productModel.productCatagory.findOneAndDelete({name:req.params.name})
     if(!catagory){
         throw new Error(`${req.params.name} is not found`)
@@ -69,7 +92,11 @@ exports.addProduct = async (req)=>{
     if(!seller){
         throw new Error('seller not found')
     }
-    const catagory = await productModel.productCatagory.findOne({name:'smartPhone'})
+    const role = ['Seller','admin']
+    if(!role.includes(seller.role)){
+        throw new Error("permission denied: Buyer can't add product")
+    }
+    const catagory = await productModel.productCatagory.findOne({name:req.body.catagory})
     if(!catagory){
         throw new Error('catagory not found')
     }
@@ -85,13 +112,19 @@ exports.addProduct = async (req)=>{
 }
 
 exports.getProduct = async (req)=>{
-    const product = await productModel.product.findOne({name:req.params.name})
-    console.log(product)
+    const user = await userModel.User.findOne({email:req.user.email})
+    const product = await productModel.product.findOne({
+        sku:req.params.sku,
+        sellerId:user._id
+
+    })
     if(!product){
         throw new Error(`${req.params.sku} is not found`)
     }
+    const catagory = await productModel.productCatagory.findOne({_id:product.catagoryId})
     return {
         name:product.name,
+        catagory:catagory.name,
         sku:product.sku,
         stockLevel:product.stockLevel,
         description:product.description
@@ -107,10 +140,13 @@ exports.updateProduct = async(req)=>{
         console.log(errorMessages)
         throw new Error(errorMessages)
     }
-    const user = await userModel.User.findOne({email:req.user.email})
-    
+    const seller = await userModel.User.findOne({email:req.user.email})
+    const role = ['Seller','admin']
+    if(!role.includes(seller.role)){
+        throw new Error("permission denied: Buyer can't add product")
+    }
     const product = await productModel.product.findOneAndUpdate(
-        {$and: [{sku:req.params.sku},{sellerId:user._id}]},
+        {$and: [{sku:req.params.sku},{sellerId:seller._id}]},
         {
             name:req.body.name,
             sku:req.body.sku,
@@ -130,16 +166,17 @@ exports.updateProduct = async(req)=>{
 
 exports.deleteProduct = async (req)=>{
 
-    const user = await userModel.User.findOne({email:req.user.email})
-    
+    const seller = await userModel.User.findOne({email:req.user.email})
+    const role = ['Seller','admin']
+    if(!role.includes(seller.role)){
+        throw new Error("permission denied: Buyer can't add product")
+    }
     const catagory = await productModel.productCatagory.deleteOne({
         sku:req.params.sku,
-        sellerId: user._id
+        sellerId: seller._id
     })
     if(!catagory){
         throw new Error(`${req.params.sku} is not found`)
     }
-    console.log(catagory)
-
     return {msg:'successfully deleted'}
 }
